+++ api/index.ts
@@ -14,6 +14,7 @@ import { setupProxies } from "./src/proxies";
 import authRouter from "./src/routes/auth";
 import configRouter from "./src/routes/config";
 import historyRouter from "./src/routes/history";
+import playRoutes from "./src/routes/play";
 import processingRouter from "./src/routes/processing";
 import sseRouter from "./src/routes/sse";
 import syncRouter from "./src/routes/sync";
@@ -72,6 +73,7 @@ app.use("/api", syncRouter);
 app.use("/api", customCssRouter);
 app.use("/api", tiddlTomlRouter);
 app.use("/api", historyRouter);
+app.use("/api", playRoutes);
 
 // Run