+++ app/src/provider/ApiFetcherProvider.tsx
@@ -65,6 +65,7 @@ type ApiFetcherContextType = {
     get_queue_status: () => Promise<{ isPaused: boolean } | undefined>;
     get_list_history: () => Promise<string[] | undefined>;
     flush_history: () => Promise<unknown>;
+    signStream: (id: string) => Promise<{ url: string } | undefined>;
   };
 };
 
@@ -448,6 +449,20 @@ export function APIFetcherProvider({ children }: { children: ReactNode }) {
     });
   }
 
+  // Play track
+
+  async function signStream(id: string) {
+    return await queryExpressJS<{ url: string }>(
+      `${apiUrl}/stream/sign/${id}`,
+      {
+        method: "GET",
+        headers: {
+          "Content-Type": "application/json",
+        },
+      },
+    );
+  }
+
   const value = {
     apiUrl,
     error: {
@@ -480,6 +495,7 @@ export function APIFetcherProvider({ children }: { children: ReactNode }) {
       get_queue_status,
       get_list_history,
       flush_history,
+      signStream,
     },
   };